Fix AMP dtype mismatch in model_c _pad_per_sample
Same fix as model.py: add .to(padded.dtype) when assigning symbol
embeddings to padded tensor. Without this, AMP training of variant C
crashes with dtype mismatch between fp16/fp32 tensors.

+++ conjecture_gen/model_c.py
@@ -142,7 +142,7 @@ class VAETransformerDecoder(nn.Module):
             sm = sorted_batch == i
             n = sm.sum().item()
             if n > 0:
-                padded[i, :n] = sorted_embeds[sm]
+                padded[i, :n] = sorted_embeds[sm].to(padded.dtype)
                 mask[i, :n] = True
         return padded, mask
